made new cart component: cart list, to access pre-loaded props and utlize component did mount for setting local storage and local state"
